Add classes for SMPTE 2094-50 -> Inject only working
