+++ IsoLib/t35_tool/SMPTE_ST2094_50.hpp
@@ -1,0 +1,148 @@
+#ifndef SMPTE_ST2094_50_HPP
+#define SMPTE_ST2094_50_HPP
+#include <iostream>
+#include <sstream>
+#include <string>
+#include <vector>
+
+// 3rd party headers
+#include <nlohmann/json.hpp>
+
+const float  PI_CUSTOM = 3.14159265358979323846;
+
+const int MAX_NB_ALTERNATE = 4;
+const int MAX_NB_CONTROL_POINTS = 32;
+const int MAX_NB_CHROMATICITIES = 8;
+const int MAX_NB_component_mixing_coefficient = 6;
+
+// Compute quantization error of each float to uint16_t
+const float Q_HDR_REFERENCE_WHITE = 50000.0 / 10000.0;
+const float Q_HDR_HEADROOM   = 60000.0 / 6.0;
+const float P_GAIN_APPLICATION_SPACE_CHROMATICITY = 3.0 / 30000.0 / 2.0;
+const float Q_GAIN_APPLICATION_SPACE_CHROMATICITY = 30000.0 / 3.0; 
+const float P_COMPONENT_MIXING_COEFFICIENT = 1.0 / 10000.0 / 2.0;
+const float Q_COMPONENT_MIXING_COEFFICIENT = 10000.0; 
+const float Q_GAIN_CURVE_CONTROL_POINT_X = 64000.0 / 64.0; 
+const float O_GAIN_CURVE_CONTROL_POINT_Y = 6.0; 
+const float Q_GAIN_CURVE_CONTROL_POINT_Y = 48000.0 / 12.0; 
+const float O_GAIN_CURVE_CONTROL_POINT_THETA = 90.0; 
+const float Q_GAIN_CURVE_CONTROL_POINT_THETA = 36000.0 / 180.0; 
+
+struct GainCurve{
+  // std::vector<uint32_t> gainCurveInterpolation; -> not in the spec currently
+  uint32_t gainCurveNumControlPoints;
+  std::vector<float> gainCurveControlPointX;
+  std::vector<float> gainCurveControlPointY;
+  std::vector<float> gainCurveControlPointTheta;
+};
+
+struct ComponentMix{
+  float componentMixRed;
+  float componentMixGreen;
+  float componentMixBlue;
+  float componentMixMax;
+  float componentMixMin;
+  float componentMixComponent;
+};
+
+struct ColorGainFunction{
+  ComponentMix cm;
+  GainCurve gc;
+};
+
+struct HeadroomAdaptiveToneMap{
+  float baselineHdrHeadroom;
+  
+  uint32_t numAlternateImages;
+  float gainApplicationSpaceChromaticities[MAX_NB_CHROMATICITIES];
+  std::vector<float> alternateHdrHeadroom;
+  std::vector<ColorGainFunction> cgf;
+};
+
+struct ColorVolumeTransform{
+  float hdrReferenceWhite;
+  HeadroomAdaptiveToneMap hatm;
+};
+
+struct ProcessingWindow{
+  uint32_t upperLeftCorner;
+  uint32_t lowerRightCorner;
+  uint32_t windowNumber;
+};
+
+struct TimeInterval{
+  uint32_t timeIntervalStart;
+  uint32_t timeintervalDuration;
+};
+
+// Structure with the syntax elements
+struct SyntaxElements {
+    // smpte_st_2094_50_application_info
+    uint16_t application_version;     
+     
+    // smpte_st_2094_50_color_volume_transform
+    bool has_custom_hdr_reference_white_flag;   
+    bool has_adaptive_tone_map_flag;                                               
+    uint16_t hdr_reference_white ;  
+    
+    // smpte_st_2094_50_adaptive_tone_map
+    uint16_t baseline_hdr_headroom ;   
+    bool use_reference_white_tone_mapping_flag;
+    uint16_t num_alternate_images;          
+    uint16_t gain_application_space_chromaticities_mode;
+    bool has_common_component_component_mixing_coefficient_flag;
+    bool has_common_curve_params_flag;
+    uint16_t gain_application_space_chromaticities[MAX_NB_CHROMATICITIES];
+    uint16_t  alternate_hdr_headrooms[MAX_NB_ALTERNATE];                
+  
+    // smpte_st_2094_50_component_mixing
+    uint16_t  component_mixing_type[MAX_NB_ALTERNATE];
+    bool has_component_mixing_coefficient_flag[MAX_NB_ALTERNATE][MAX_NB_component_mixing_coefficient];
+    uint16_t  component_mixing_coefficient[MAX_NB_ALTERNATE][MAX_NB_component_mixing_coefficient];
+  
+    // smpte_st_2094_50_gain_curve
+    uint16_t  gain_curve_num_control_points_minus_1[MAX_NB_ALTERNATE];
+    bool  gain_curve_use_pchip_slope_flag[MAX_NB_ALTERNATE];
+    uint16_t  gain_curve_control_points_x[MAX_NB_ALTERNATE][MAX_NB_CONTROL_POINTS];         
+    uint16_t  gain_curve_control_points_y[MAX_NB_ALTERNATE][MAX_NB_CONTROL_POINTS];            
+    uint16_t  gain_curve_control_points_theta[MAX_NB_ALTERNATE][MAX_NB_CONTROL_POINTS];     
+  };
+
+class SMPTE_ST2094_50 {
+public:
+    SMPTE_ST2094_50(); // Constructor
+    bool decodeJsonToMetadataItems(nlohmann::json j, const std::string& path); 
+    void convertMetadataItemsToSyntaxElements();
+    void writeSyntaxElementsToBinaryData();
+
+    //void SMPTE_ST2094_50::decodeBinaryToSyntaxElements(std::vector<uint8_t> binary_data) ;
+    //void SMPTE_ST2094_50::convertSyntaxElementsToMetadataItems();
+
+    // Getters
+    std::vector<uint8_t>    getPayloadData();
+    uint32_t                getTimeIntervalStart();
+    uint32_t                getTimeintervalDuration();
+
+
+    void dbgPrintMetadataItems(bool decode);
+    // Carrying mechanism information
+    std::string keyValue;
+    std::vector<uint8_t> payloadBinaryData;
+
+private:
+    uint8_t applicationIdentifier;
+    uint8_t applicationVersion;
+    TimeInterval timeI;
+    ProcessingWindow pWin;
+    ColorVolumeTransform cvt;
+
+    // not in specification, convenience flag for implementation
+    bool isHeadroomAdaptiveToneMap;   
+    bool isReferenceWhiteToneMapping;
+    bool isLinearInterpolation[MAX_NB_ALTERNATE];
+    bool hasSlopeParameter[MAX_NB_ALTERNATE];
+
+    SyntaxElements elm;
+};
+
+#endif // SMPTE_ST2094_50_HPP
